refactor universal channel bool in config and using config as env variable
