+++ Justfile
@@ -24,39 +24,34 @@ compile COMPILER='hardhat':
 # Usage: just set-contracts [chain] [contract-type]
 set-contracts CHAIN CONTRACT_TYPE:
     echo "Updating config.json with contract type..."
-    node scripts/set-contracts-config.js {{CHAIN}} {{CONTRACT_TYPE}}
+    node scripts/_set-contracts-config.js {{CHAIN}} {{CONTRACT_TYPE}}
 
 # Deploy the contracts in the /contracts folder using Hardhat and updating the config.json file
 # The source and destination arguments are REQUIRED;
 # The universal argument is optional; if not provided, it defaults to "true".
-# Usage: just deploy [source] [destination] [config-file]
+# Usage: just deploy [source] [destination]
 deploy SOURCE DESTINATION:
         echo "Deploying contracts with Hardhat..."
-        node scripts/deploy-config.js {{SOURCE}} {{DESTINATION}}
+        node scripts/_deploy-config.js {{SOURCE}} {{DESTINATION}}
 
 # Run the sanity check script to verify that configuration (.env) files match with deployed contracts' stored values
 # Usage: just sanity-check
 sanity-check:
     echo "Running sanity check..."
-    node scripts/sanity-check.js
+    node scripts/_sanity-check.js
+# TODO: add a check to see if isUniversal flag is correct
 
-# Update the universal channel handler address in the on the IBC contract with that from the .env file
-# Usage: just update-uc-handler [chain]
-update-uc-handler CHAIN:
-  echo "Updating the universal channel handler address..."
-  npx hardhat run scripts/_updateUcHandler.js --network {{CHAIN}}
-
-# Update the dispatcher address in the on the IBC contract with that from the .env file
-# Usage: just dispatcher [chain]
-update-dispatcher CHAIN:
-  echo "Updating the universal channel handler address..."
-  npx hardhat run scripts/_updateDispatcher.js --network {{CHAIN}}
+# Update the dispatcher or universal channel handler address on the IBC application, with that from the .env file
+# Usage: just update-vibc [chain]
+update-vibc CHAIN:
+  echo "Updating the dispatcher or universal channel handler address..."
+  npx hardhat run scripts/_update-vibc-address.js --network {{CHAIN}}
 
 # Create a channel by triggering a channel handshake from the source and with parameters found in the config.json file
 # Usage: just create-channel
 create-channel:
     echo "Attempting to create a channel with the values from the config..."
-    node scripts/create-channel-config.js
+    node scripts/_create-channel-config.js
 
 # Send a packet over the universal channel or a custom channel as defined in the config.json file
 # The source argument is REQUIRED;
@@ -64,33 +59,25 @@ create-channel:
 # Usage: just send-packet [source] [universal]
 send-packet SOURCE:
     echo "Sending a packet with the values from the config..."
-    node scripts/send-packet-config.js {{SOURCE}}
+    node scripts/_send-packet-config.js {{SOURCE}}
 
-# Usage: just switch-client [universal=true]
-switch-client UNIVERSAL='true':
-    #!/usr/bin/env sh
+# Switch between the sim client and the client with proofs
+# Usage: just switch-client
+switch-client:
     echo "Switching between sim client and client with proofs..."
-    node scripts/switch-clients.js
-    if test "{{UNIVERSAL}}" = "true"; then
-        npx hardhat run scripts/_updateUcHandler.js --network optimism
-        npx hardhat run scripts/_updateUcHandler.js --network base
-    elif test "{{UNIVERSAL}}" = "false"; then
-        npx hardhat run scripts/_updateDispatcher.js --network optimism
-        npx hardhat run scripts/_updateDispatcher.js --network base
-    else
-        echo "Unknown universal flag: {{UNIVERSAL}}"
-        exit 1
-    fi
-
+    npx hardhat run scripts/_update-vibc-address.js --network optimism
+    npx hardhat run scripts/_update-vibc-address.js --network base
+    node scripts/_switch-clients.js
 
 # Run the full E2E flow by setting the contracts, deploying them, creating a channel, and sending a packet
 # Usage: just do-it
 do-it:
     echo "Running the full E2E flow..."
     just set-contracts optimism XCounter && just set-contracts base XCounter
-    just deploy optimism base false
+    just deploy optimism base
+    just sanity-check
     just create-channel
-    just send-packet optimism false
+    just send-packet optimism
     echo "You've done it!"
 
 # Clean up the environment by removing the artifacts and cache folders and running the forge clean command
