+++ scripts/_sanity-check.js
@@ -5,10 +5,9 @@ function runSanityCheck(network) {
     // Example: 
     // $ node scripts/sanity-check.js
     const configPath = getConfigPath();
-
     const config = require(configPath);
 
-    const scriptSuffix = config.isUniversalChannel === 'true' ? 'universal' : 'custom';
+    const scriptSuffix = config.isUniversal ? 'universal' : 'custom';
     exec(`npx hardhat run scripts/_sanity-check-${scriptSuffix}.js --network ${network}`, (error, stdout, stderr) => {
         if (error) {
         console.error(`exec error: ${error}`);
