+++ scripts/_set-contracts-config.js
