+++ scripts/send-universal-packet.js
@@ -5,23 +5,14 @@
 // will compile your contracts, add the Hardhat Runtime Environment's members to the
 // global scope, and execute the script.
 const hre = require('hardhat');
-const path = require('path');
-const configRelativePath = process.env.CONFIG_PATH || 'config.json';
-const configPath = path.join(__dirname, '..' , configRelativePath);
-const config = require(configPath);
-const sendConfig = config.sendUniversalPacket;
-
-const { getIbcApp, getDispatcher } = require('./_vibc-helpers.js');
-const { listenForIbcPacketEvents } = require('./_events.js');
+const { getConfigPath } = require('./_helpers');
+const { getIbcApp } = require('./_vibc-helpers.js');
 
 async function main() {
     const accounts = await hre.ethers.getSigners();
-
-    // Get the dispatchers for both source and destination to listen for IBC packet events
-    const opDispatcher = await getDispatcher("optimism");
-    const baseDispatcher = await getDispatcher("base");
-    listenForIbcPacketEvents("optimism", opDispatcher);
-    listenForIbcPacketEvents("base", baseDispatcher);
+    const configPath = getConfigPath();
+    const config = require(configPath);
+    const sendConfig = config.sendUniversalPacket;
 
     const networkName = hre.network.name;
     // Get the contract type from the config and get the contract
@@ -30,7 +21,9 @@ async function main() {
     // Do logic to prepare the packet
 
     // If the network we are sending on is optimism, we need to use the base port address and vice versa
-    const destPortAddr = networkName === "optimism" ? config["sendUniversalPacket"]["base"]["portAddr"] : config["sendUniversalPacket"]["optimism"]["portAddr"];
+    const destPortAddr = networkName === "optimism" ?
+      config["sendUniversalPacket"]["base"]["portAddr"] :
+      config["sendUniversalPacket"]["optimism"]["portAddr"];
     const channelId = sendConfig[`${networkName}`]["channelId"];
     const channelIdBytes = hre.ethers.encodeBytes32String(channelId);
     const timeoutSeconds = sendConfig[`${networkName}`]["timeout"];
