--- scripts/test.js
@@ -1,4 +1,0 @@
-const { getConfigPath } = require('./_helpers.js');
-
-const configPath = getConfigPath('config-universal');
-console.log(configPath);
